+++ .github/workflows/reanimated-worklets-compatibility-check.yml
@@ -7,11 +7,13 @@ on:
     paths:
       - '.github/workflows/reanimated-worklets-compatibility-check.yml'
       - '.github/workflows/helper/read-reanimated-worklets-compatibility.ts'
+      - '.github/workflows/helper/npm-versions.ts'
       - '.github/actions/create-rnc-cli-app/**'
   pull_request:
     paths:
       - '.github/workflows/reanimated-worklets-compatibility-check.yml'
       - '.github/workflows/helper/read-reanimated-worklets-compatibility.ts'
+      - '.github/workflows/helper/npm-versions.ts'
       - '.github/actions/create-rnc-cli-app/**'
   workflow_dispatch:
   schedule:
